[rp2040] Move HAL bodies into components/rp2040/hal.cpp + inline trivial dispatches
Per-platform follow-up to #15977.

Move out-of-line HAL bodies (arch_init, arch_restart, arch_get_cpu_freq_hz)
from components/rp2040/core.cpp into a new components/rp2040/hal.cpp.
arch_init keeps its full body (crash handler init + watchdog enable) since
it depends on the component-private crash_handler.h. arch_get_cpu_freq_hz
calls RP2040::f_cpu() (Arduino-pico C++ class) so it stays out-of-line to
avoid pulling that header into hal_rp2040.h. core.cpp becomes an empty
stub.

Inline the trivial one-liners directly in hal_rp2040.h with forward decls:

  delayMicroseconds(us)        -> delay_microseconds_safe(us)
  arch_feed_wdt()              -> watchdog_update()
  arch_get_cpu_cycle_count()   -> ulMainGetRunTimeCounterValue()
